# faster path filling again